refactor(compile): extract eval-worker materialization into a submodule
collect_modules.rs was at the 2000-line file-size gate; move the eval-mode
Worker source materialization helper + its test into collect_modules/eval_worker.rs
(mirrors the existing crypto_ns / dynamic_glob / wasm_asset splits).
